# sd: enable 4 bit bus width for high speed cards