Enable 4 bit bus width for high speed cards, so that host and card combinations that cannot use UHS mode will still benefit from the speed increase of using 4 data lines.


Some performance results using a Kingston 2GB SD card that only supports high speed mode (collected on an LPCXpresso55s69):

Without 4 bit bus mode:
| Read | Write |
| ---- | ---- |
| 3504 KiB/s | 2592 KiB/s |

With 4 bit bus mode:
| Read | Write |
| ---- | ---- |
| 9840 KiB/s | 5280 KiB/s |